refactor: Separate GameData preprocessing from Weapon and Accessory class
