+++ Assets/Script/Equipment/Accessory/Accessory.cs
@@ -9,20 +9,12 @@ public class Accessory
     public int AccessoryLevel;
     public int AccessoryMaxLevel;
 
-    public static List<List<Tuple<int, float>>>[] AccessoryUpgrade;
-    public static int[] AccessoriesMaxLevel;
-
-    // Start is called before the first frame update
-    static Accessory()
-    {
-        AccessoryUpgradePreprocessing();
-    }
-
+    public EquipmentData EquipmentData;
     public Accessory(int accessoryIndex)
     {
         AccessoryIndex = accessoryIndex;
         AccessoryLevel = 0;
-        AccessoryMaxLevel = AccessoriesMaxLevel[accessoryIndex];
+        AccessoryMaxLevel = EquipmentData.AccessoriesMaxLevel[accessoryIndex];
     }
 
     public bool IsMaster()
@@ -33,693 +25,13 @@ public class Accessory
     {
         Character character = GameObject.Find("Player").GetComponent<Character>();
         AccessoryLevel++;
-        foreach ((var statIndex, var data) in AccessoryUpgrade[AccessoryIndex][AccessoryLevel])
+        foreach ((var statIndex, var data) in EquipmentData.AccessoryUpgrade[AccessoryIndex][AccessoryLevel])
         {
             character.CharacterStats[statIndex] += data;
             if (statIndex == (int)Enums.Stat.Luck)
             {
                 character.UpdateLuck(character.CharacterStats[statIndex]);
             }
-        }
-    }
-    private static void AccessoryUpgradePreprocessing()
-    {
-        AccessoryUpgrade = new List<List<Tuple<int, float>>>[21];
-        // 앞의 한 개(index상 0에 해당)는 더미 데이터
-        // Spinach
-        AccessoryUpgrade[0] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[0].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[0].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.1f)
-        });
-        AccessoryUpgrade[0].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.1f)
-        });
-        AccessoryUpgrade[0].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.1f)
-        });
-        AccessoryUpgrade[0].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.1f)
-        });
-        AccessoryUpgrade[0].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.1f)
-        });
-
-        // Armor
-        AccessoryUpgrade[1] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[1].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[1].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Armor, 1),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ReflectionPer, 0.1f)
-        });
-        AccessoryUpgrade[1].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Armor, 1),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ReflectionPer, 0.1f)
-        });
-        AccessoryUpgrade[1].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Armor, 1),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ReflectionPer, 0.1f)
-        });
-        AccessoryUpgrade[1].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Armor, 1),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ReflectionPer, 0.1f)
-        });
-        AccessoryUpgrade[1].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Armor, 1),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ReflectionPer, 0.1f)
-        });
-
-        // HollowHeart
-        AccessoryUpgrade[2] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[2].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[2].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.2f)
-        });
-        AccessoryUpgrade[2].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.2f)
-        });
-        AccessoryUpgrade[2].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.2f)
-        });
-        AccessoryUpgrade[2].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.2f)
-        });
-        AccessoryUpgrade[2].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.2f)
-        });
-
-        // Pummarola
-        AccessoryUpgrade[3] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[3].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[3].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f)
-        });
-        AccessoryUpgrade[3].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f)
-        });
-        AccessoryUpgrade[3].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f)
-        });
-        AccessoryUpgrade[3].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f)
-        });
-        AccessoryUpgrade[3].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f)
-        });
-
-        // EmptyTome
-        AccessoryUpgrade[4] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[4].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[4].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CooldownPer, -0.8f)
-        });
-        AccessoryUpgrade[4].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CooldownPer, -0.8f)
-        });
-        AccessoryUpgrade[4].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CooldownPer, -0.8f)
-        });
-        AccessoryUpgrade[4].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CooldownPer, -0.8f)
-        });
-        AccessoryUpgrade[4].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CooldownPer, -0.8f)
-        });
-
-        // Candelabrador
-        AccessoryUpgrade[5] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[5].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[5].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.10f)
-        });
-        AccessoryUpgrade[5].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.10f)
-        });
-        AccessoryUpgrade[5].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.10f)
-        });
-        AccessoryUpgrade[5].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.10f)
-        });
-        AccessoryUpgrade[5].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.10f)
-        });
-
-        // Bracer
-        AccessoryUpgrade[6] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[6].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[6].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.1f)
-        });
-        AccessoryUpgrade[6].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.1f)
-        });
-        AccessoryUpgrade[6].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.1f)
-        });
-        AccessoryUpgrade[6].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.1f)
-        });
-        AccessoryUpgrade[6].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.1f)
-        });
-
-        // Spellbinder
-        AccessoryUpgrade[7] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[7].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[7].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.1f)
-        });
-        AccessoryUpgrade[7].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.1f)
-        });
-        AccessoryUpgrade[7].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.1f)
-        });
-        AccessoryUpgrade[7].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.1f)
-        });
-        AccessoryUpgrade[7].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.1f)
-        });
-
-        // Duplicator
-        AccessoryUpgrade[8] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[8].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[8].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Amount, 1)
-        });
-        AccessoryUpgrade[8].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Amount, 1)
-        });
-
-        // Wings
-        AccessoryUpgrade[9] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[9].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[9].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MoveSpeedPer, 10)
-        });
-        AccessoryUpgrade[9].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MoveSpeedPer, 10)
-        });
-        AccessoryUpgrade[9].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MoveSpeedPer, 10)
-        });
-        AccessoryUpgrade[9].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MoveSpeedPer, 10)
-        });
-        AccessoryUpgrade[9].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MoveSpeedPer, 10)
-        });
-
-        // Attractorb
-        AccessoryUpgrade[10] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[10].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[10].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MagnetPer, 33)
-        });
-        AccessoryUpgrade[10].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MagnetPer, 33)
-        });
-        AccessoryUpgrade[10].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MagnetPer, 25)
-        });
-        AccessoryUpgrade[10].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MagnetPer, 20)
-        });
-        AccessoryUpgrade[10].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MagnetPer, 33)
-        });
-
-        // Clover
-        AccessoryUpgrade[11] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[11].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[11].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.LuckPer, 10)
-        });
-        AccessoryUpgrade[11].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.LuckPer, 10)
-        });
-        AccessoryUpgrade[11].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.LuckPer, 10)
-        });
-        AccessoryUpgrade[11].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.LuckPer, 10)
-        });
-        AccessoryUpgrade[11].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.LuckPer, 10)
-        });
-
-        // Crown
-        AccessoryUpgrade[12] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[12].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[12].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GrowthPer, 8)
-        });
-        AccessoryUpgrade[12].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GrowthPer, 8)
-        });
-        AccessoryUpgrade[12].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GrowthPer, 8)
-        });
-        AccessoryUpgrade[12].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GrowthPer, 8)
-        });
-        AccessoryUpgrade[12].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GrowthPer, 8)
-        });
-
-        // StoneMask
-        AccessoryUpgrade[13] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[13].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[13].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GreedPer, 10)
-        });
-        AccessoryUpgrade[13].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GreedPer, 10)
-        });
-        AccessoryUpgrade[13].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GreedPer, 10)
-        });
-        AccessoryUpgrade[13].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GreedPer, 10)
-        });
-        AccessoryUpgrade[13].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.GreedPer, 10)
-        });
-
-        // Tiragisu
-        AccessoryUpgrade[14] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[14].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[14].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Revival, 1)
-        });
-        AccessoryUpgrade[14].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Revival, 1)
-        });
-
-        // Skull
-        AccessoryUpgrade[15] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[15].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[15].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 10)
-        });
-        AccessoryUpgrade[15].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 10)
-        });
-        AccessoryUpgrade[15].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 10)
-        });
-        AccessoryUpgrade[15].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 10)
-        });
-        AccessoryUpgrade[15].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 10)
-        });
-
-        // SilverRing
-        AccessoryUpgrade[16] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-        AccessoryUpgrade[16].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.05f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.05f)
-        });
-
-        // GoldRing
-        AccessoryUpgrade[17] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[17].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-
-        // MetaglioLeft
-        AccessoryUpgrade[18] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-        AccessoryUpgrade[18].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.Recovery, 0.1f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.MaxHealthPer, 0.05f)
-        });
-
-        // MetaglioRight
-        AccessoryUpgrade[19] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-        AccessoryUpgrade[19].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 5)
-        });
-
-        // TorronaBox
-        AccessoryUpgrade[20] = new List<List<Tuple<int, float>>>();
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>()
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.0f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.03f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.03f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.03f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.03f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.03f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.03f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.03f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.MightPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.ProjectileSpeedPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.DurationPer, 0.03f),
-            new Tuple<int, float>((int)Enums.AccessoryStat.AreaPer, 0.03f)
-        });
-        AccessoryUpgrade[20].Add(new List<Tuple<int, float>>
-        {
-            new Tuple<int, float>((int)Enums.AccessoryStat.CursePer, 100)
-        });
-
-        AccessoriesMaxLevel = new int[21];
-        for (int i = 0; i < 21; i++)
-        {
-            AccessoriesMaxLevel[i] = AccessoryUpgrade[i].Count - 1;
         }
     }
 }
