+++ Assets/Script/GameData.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1cc24a5d27060394ca1480d424c6db03
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
