+++ Assets/Script/GameData/EquipmentData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1f6b86f0b7ff0974e90866815dcb9250
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
